refactor: split core services and models into focused extensions
