+++ solidarity/Models/BusinessCard.swift
@@ -7,7 +7,6 @@
 
 import Foundation
 
-/// Main business card data model with comprehensive contact information and privacy controls
 struct BusinessCard: Codable, Identifiable, Equatable, Hashable {
   let id: UUID
   var name: String
@@ -21,7 +20,10 @@ struct BusinessCard: Codable, Identifiable, Equatable, Hashable {
   var skills: [Skill]
   var categories: [String]
   var sharingPreferences: SharingPreferences
-  var groupContext: GroupCredentialContext?  // Group VC Context
+  var groupContext: GroupCredentialContext?
+  /// Fields that have been cryptographically verified (e.g., from passport, signed exchange).
+  /// nil means no verification tracking (legacy cards); empty set means nothing verified.
+  var verifiedFields: Set<BusinessCardField>?
   var createdAt: Date
   var updatedAt: Date
 
@@ -39,6 +41,7 @@ struct BusinessCard: Codable, Identifiable, Equatable, Hashable {
     categories: [String] = [],
     sharingPreferences: SharingPreferences = SharingPreferences(),
     groupContext: GroupCredentialContext? = nil,
+    verifiedFields: Set<BusinessCardField>? = nil,
     createdAt: Date = Date(),
     updatedAt: Date = Date()
   ) {
@@ -55,16 +58,15 @@ struct BusinessCard: Codable, Identifiable, Equatable, Hashable {
     self.categories = categories
     self.sharingPreferences = sharingPreferences
     self.groupContext = groupContext
+    self.verifiedFields = verifiedFields
     self.createdAt = createdAt
     self.updatedAt = updatedAt
   }
 
-  /// Update the business card and refresh the updatedAt timestamp
   mutating func update() {
     self.updatedAt = Date()
   }
 
-  /// Get filtered business card based on sharing preferences for a specific sharing level
   func filteredCard(for sharingLevel: SharingLevel) -> BusinessCard {
     var filtered = self
     let allowedFields = sharingPreferences.effectiveFields(preferredLevel: sharingLevel)
@@ -78,13 +80,9 @@ struct BusinessCard: Codable, Identifiable, Equatable, Hashable {
     if !allowedFields.contains(.socialNetworks) { filtered.socialNetworks = [] }
     if !allowedFields.contains(.skills) { filtered.skills = [] }
 
-    // If sharing level is personal/professional, we might want to strip group context
-    // unless explicitly handled. For now, we keep it as it's part of the credential.
-
     return filtered
   }
 
-  /// Get filtered business card based on explicit field set
   func filteredCard(for fields: Set<BusinessCardField>) -> BusinessCard {
     var filtered = self
     if !fields.contains(.name) { filtered.name = "" }
@@ -97,9 +95,18 @@ struct BusinessCard: Codable, Identifiable, Equatable, Hashable {
     if !fields.contains(.skills) { filtered.skills = [] }
     return filtered
   }
+
+  /// Returns a copy containing only fields that have been cryptographically verified.
+  /// Name is always included as the minimum identity field.
+  func filteredCardForVerifiedOnly() -> BusinessCard {
+    let verified = verifiedFields ?? []
+    // Name is always included as minimum identity
+    var fields = verified
+    fields.insert(.name)
+    return filteredCard(for: fields)
+  }
 }
 
-/// Social network information
 struct SocialNetwork: Codable, Identifiable, Equatable, Hashable {
   let id: UUID
   var platform: SocialPlatform
@@ -119,7 +126,6 @@ struct SocialNetwork: Codable, Identifiable, Equatable, Hashable {
   }
 }
 
-/// Supported social media platforms
 enum SocialPlatform: String, Codable, CaseIterable {
   case linkedin = "LinkedIn"
   case twitter = "Twitter"
@@ -142,7 +148,6 @@ enum SocialPlatform: String, Codable, CaseIterable {
   }
 }
 
-/// Individual skill with categorization and proficiency levels
 struct Skill: Codable, Identifiable, Equatable, Hashable {
   let id: UUID
   var name: String
@@ -162,7 +167,6 @@ struct Skill: Codable, Identifiable, Equatable, Hashable {
   }
 }
 
-/// Proficiency levels for skills
 enum ProficiencyLevel: String, Codable, CaseIterable {
   case beginner = "Beginner"
   case intermediate = "Intermediate"
@@ -179,7 +183,6 @@ enum ProficiencyLevel: String, Codable, CaseIterable {
   }
 }
 
-/// Privacy controls for selective information sharing
 struct SharingPreferences: Codable, Equatable, Hashable {
   var publicFields: Set<BusinessCardField>
   var professionalFields: Set<BusinessCardField>
@@ -198,7 +201,6 @@ struct SharingPreferences: Codable, Equatable, Hashable {
     useZK: Bool = true,
     sharingFormat: SharingFormat = .plaintext
   ) {
-    // Ensure name is always included in all levels
     var publicSet = publicFields
     publicSet.insert(.name)
     self.publicFields = publicSet
@@ -217,7 +219,6 @@ struct SharingPreferences: Codable, Equatable, Hashable {
     self.sharingFormat = sharingFormat
   }
 
-  /// Get allowed fields for a specific sharing level
   func fieldsForLevel(_ level: SharingLevel) -> Set<BusinessCardField> {
     switch level {
     case .`public`:
@@ -229,7 +230,6 @@ struct SharingPreferences: Codable, Equatable, Hashable {
     }
   }
 
-  /// Field toggles are the canonical source when all legacy levels carry the same set.
   func effectiveFields(preferredLevel: SharingLevel) -> Set<BusinessCardField> {
     if publicFields == professionalFields && professionalFields == personalFields {
       return publicFields
@@ -238,51 +238,6 @@ struct SharingPreferences: Codable, Equatable, Hashable {
   }
 }
 
-// MARK: - Codable compatibility for SharingPreferences (handle missing keys)
-
-extension SharingPreferences {
-  private enum CodingKeys: String, CodingKey {
-    case publicFields, professionalFields, personalFields, allowForwarding, expirationDate, useZK, sharingFormat
-  }
-
-  init(from decoder: Decoder) throws {
-    let container = try decoder.container(keyedBy: CodingKeys.self)
-    var publicSet =
-      try container.decodeIfPresent(Set<BusinessCardField>.self, forKey: .publicFields) ?? [.name, .title, .company]
-    publicSet.insert(.name)  // Ensure name is always included
-    self.publicFields = publicSet
-
-    var professionalSet =
-      try container.decodeIfPresent(Set<BusinessCardField>.self, forKey: .professionalFields)
-      ?? [.name, .title, .company, .email, .skills]
-    professionalSet.insert(.name)  // Ensure name is always included
-    self.professionalFields = professionalSet
-
-    var personalSet =
-      try container.decodeIfPresent(Set<BusinessCardField>.self, forKey: .personalFields)
-      ?? BusinessCardField.allCases.asSet()
-    personalSet.insert(.name)  // Ensure name is always included
-    self.personalFields = personalSet
-
-    self.allowForwarding = try container.decodeIfPresent(Bool.self, forKey: .allowForwarding) ?? false
-    self.expirationDate = try container.decodeIfPresent(Date.self, forKey: .expirationDate)
-    self.useZK = try container.decodeIfPresent(Bool.self, forKey: .useZK) ?? true
-    self.sharingFormat = try container.decodeIfPresent(SharingFormat.self, forKey: .sharingFormat) ?? .plaintext
-  }
-
-  func encode(to encoder: Encoder) throws {
-    var container = encoder.container(keyedBy: CodingKeys.self)
-    try container.encode(publicFields, forKey: .publicFields)
-    try container.encode(professionalFields, forKey: .professionalFields)
-    try container.encode(personalFields, forKey: .personalFields)
-    try container.encode(allowForwarding, forKey: .allowForwarding)
-    try container.encodeIfPresent(expirationDate, forKey: .expirationDate)
-    try container.encode(useZK, forKey: .useZK)
-    try container.encode(sharingFormat, forKey: .sharingFormat)
-  }
-}
-
-/// Available business card fields for privacy control
 enum BusinessCardField: String, Codable, CaseIterable {
   case name
   case title
@@ -320,8 +275,6 @@ enum BusinessCardField: String, Codable, CaseIterable {
   }
 }
 
-/// Legacy sharing-level marker kept for backward compatibility.
-/// Field toggles (`BusinessCardField`) are the canonical sharing controls.
 enum SharingLevel: String, Codable, CaseIterable {
   case `public` = "public"
   case professional = "professional"
@@ -349,92 +302,5 @@ enum SharingLevel: String, Codable, CaseIterable {
     case .professional: return "For work contacts and events. Usually includes email and skills."
     case .personal: return "For close contacts. Typically includes all fields."
     }
-  }
-}
-
-// MARK: - Extensions
-
-extension Array where Element == BusinessCardField {
-  func asSet() -> Set<BusinessCardField> {
-    return Set(self)
-  }
-}
-
-extension BusinessCardField: Identifiable {
-  var id: String { self.rawValue }
-}
-
-extension SharingLevel: Identifiable {
-  var id: String { self.rawValue }
-}
-
-// MARK: - Additional Extensions for Contact Management
-
-extension BusinessCard {
-  /// Get initials for profile display
-  var initials: String {
-    let components = name.components(separatedBy: " ")
-    let initials = components.compactMap { $0.first }.map { String($0) }
-    return initials.prefix(2).joined().uppercased()
-  }
-
-  /// Get profile image URL if stored as URL string
-  var profileImageURL: URL? {
-    // In a real implementation, this might return a URL to a stored image
-    // For now, return nil as we're storing image data directly
-    return nil
-  }
-
-  /// Generate vCard data for sharing
-  var vCardData: String {
-    var vCard = "BEGIN:VCARD\n"
-    vCard += "VERSION:3.0\n"
-    vCard += "FN:\(name)\n"
-
-    if let title = title, !title.isEmpty {
-      vCard += "TITLE:\(title)\n"
-    }
-
-    if let company = company, !company.isEmpty {
-      vCard += "ORG:\(company)\n"
-    }
-
-    if let email = email, !email.isEmpty {
-      vCard += "EMAIL:\(email)\n"
-    }
-
-    if let phone = phone, !phone.isEmpty {
-      vCard += "TEL:\(phone)\n"
-    }
-
-    // Add social networks
-    for social in socialNetworks {
-      if let url = social.url, !url.isEmpty {
-        vCard += "URL:\(url)\n"
-      }
-    }
-
-    // Add skills as notes
-    if !skills.isEmpty {
-      let skillsText = skills.map { "\($0.name) (\($0.proficiencyLevel.rawValue))" }.joined(separator: ", ")
-      vCard += "NOTE:Skills: \(skillsText)\n"
-    }
-
-    vCard += "END:VCARD\n"
-    return vCard
-  }
-
-  /// Sample business card for previews
-  static var sample: BusinessCard {
-    return BusinessCard(
-      name: "Solidarity User",
-      title: "Builder",
-      company: "Solidarity",
-      email: "hello@solidarity.id",
-      phone: "",
-      socialNetworks: [],
-      skills: [],
-      categories: []
-    )
   }
 }
