+++ solidarity/Services/Utils/KeyManager.swift
@@ -10,21 +10,18 @@ import CryptoKit
 import Foundation
 import Security
 
-/// Manages cryptographic keys for ZK-proof system and domain verification
 class KeyManager {
   static let shared = KeyManager()
 
-  private let keychain = KeychainManager()
+  let keychain = KeychainManager()
 
-  // Key identifiers
-  private enum KeyIdentifier: String, CaseIterable {
+  enum KeyIdentifier: String, CaseIterable {
     case masterKey = "master_key"
     case signingKey = "signing_key"
     case verificationKey = "verification_key"
     case domainKey = "domain_key"
     case proofKey = "proof_key"
 
-    // Keep the old namespace readable for migration and downgrade safety.
     var legacyKeychainTag: String {
       return "com.kidneyweakx.airmeishi.keys.\(self.rawValue)"
     }
@@ -58,21 +55,17 @@ class KeyManager {
 
   // MARK: - Public Methods
 
-  /// Initialize key management system
   func initializeKeys() -> CardResult<Void> {
-    // Generate master key if it doesn't exist
     let masterKeyResult = getOrCreateMasterKey()
     guard case .success = masterKeyResult else {
       return .failure(.encryptionError("Failed to initialize master key"))
     }
 
-    // Generate signing key pair if it doesn't exist
     let signingKeyResult = getOrCreateSigningKeyPair()
     guard case .success = signingKeyResult else {
       return .failure(.encryptionError("Failed to initialize signing keys"))
     }
 
-    // Generate domain verification key if it doesn't exist
     let domainKeyResult = getOrCreateDomainKey()
     guard case .success = domainKeyResult else {
       return .failure(.encryptionError("Failed to initialize domain key"))
@@ -81,27 +74,22 @@ class KeyManager {
     return .success(())
   }
 
-  /// Get master key for general encryption
   func getMasterKey() -> CardResult<SymmetricKey> {
     return getOrCreateMasterKey()
   }
 
-  /// Get signing key pair for digital signatures
   func getSigningKeyPair() -> CardResult<(privateKey: P256.Signing.PrivateKey, publicKey: P256.Signing.PublicKey)> {
     return getOrCreateSigningKeyPair()
   }
 
-  /// Get domain verification key
   func getDomainKey() -> CardResult<SymmetricKey> {
     return getOrCreateDomainKey()
   }
 
-  /// Generate ephemeral key for one-time use
   func generateEphemeralKey() -> SymmetricKey {
     return SymmetricKey(size: .bits256)
   }
 
-  /// Generate key derivation for specific purpose
   func deriveKey(from masterKey: SymmetricKey, purpose: String, context: String) -> CardResult<SymmetricKey> {
     let info = Data("\(purpose):\(context)".utf8)
 
@@ -114,20 +102,16 @@ class KeyManager {
     return .success(derivedKey)
   }
 
-  /// Rotate all keys (for security maintenance)
   func rotateKeys() -> CardResult<Void> {
-    // Delete existing keys
     for keyId in KeyIdentifier.allCases {
       for tag in keyId.allTags {
         _ = keychain.deleteKey(tag: tag)
       }
     }
 
-    // Reinitialize with new keys
     return initializeKeys()
   }
 
-  /// Export public keys for sharing
   func exportPublicKeys() -> CardResult<PublicKeyBundle> {
     let signingKeyResult = getSigningKeyPair()
 
@@ -149,14 +133,11 @@ class KeyManager {
     }
   }
 
-  /// Verify a public key bundle
   func verifyPublicKeyBundle(_ bundle: PublicKeyBundle) -> CardResult<Bool> {
-    // Check expiration
     if bundle.expiresAt < Date() {
       return .success(false)
     }
 
-    // Verify key format
     do {
       _ = try P256.Signing.PublicKey(rawRepresentation: bundle.signingPublicKey)
       return .success(true)
@@ -165,7 +146,6 @@ class KeyManager {
     }
   }
 
-  /// Clear all keys (for app reset)
   func clearAllKeys() -> CardResult<Void> {
     var hasError = false
 
@@ -192,7 +172,6 @@ class KeyManager {
       return .success(key)
 
     case .failure:
-      // Generate new key
       let newKey = SymmetricKey(size: .bits256)
       let storeResult = storeSymmetricKey(newKey, identifier: identifier)
 
@@ -216,7 +195,6 @@ class KeyManager {
       return .success((privateKey: privateKey, publicKey: privateKey.publicKey))
 
     case .failure:
-      // Generate new key pair
       let newPrivateKey = P256.Signing.PrivateKey()
       let storeResult = storePrivateKey(newPrivateKey, identifier: identifier)
 
@@ -238,7 +216,6 @@ class KeyManager {
       return .success(key)
 
     case .failure:
-      // Generate new key
       let newKey = SymmetricKey(size: .bits256)
       let storeResult = storeSymmetricKey(newKey, identifier: identifier)
 
@@ -251,174 +228,4 @@ class KeyManager {
     }
   }
 
-  private func retrieveSymmetricKey(_ identifier: KeyIdentifier) -> CardResult<SymmetricKey> {
-    for tag in identifier.readTags {
-      let retrieve = keychain.retrieveSymmetricKey(tag: tag)
-      if case .success(let key) = retrieve {
-        if tag != identifier.primaryTag {
-          _ = keychain.storeSymmetricKey(key, tag: identifier.primaryTag)
-        }
-        if tag != identifier.legacyKeychainTag {
-          _ = keychain.storeSymmetricKey(key, tag: identifier.legacyKeychainTag)
-        }
-        return .success(key)
-      }
-    }
-    return .failure(.encryptionError("Failed to retrieve symmetric key from compatible tags"))
-  }
-
-  private func retrievePrivateKey(_ identifier: KeyIdentifier) -> CardResult<P256.Signing.PrivateKey> {
-    for tag in identifier.readTags {
-      let retrieve = keychain.retrievePrivateKey(tag: tag)
-      if case .success(let key) = retrieve {
-        if tag != identifier.primaryTag {
-          _ = keychain.storePrivateKey(key, tag: identifier.primaryTag)
-        }
-        if tag != identifier.legacyKeychainTag {
-          _ = keychain.storePrivateKey(key, tag: identifier.legacyKeychainTag)
-        }
-        return .success(key)
-      }
-    }
-    return .failure(.encryptionError("Failed to retrieve private key from compatible tags"))
-  }
-
-  private func storeSymmetricKey(_ key: SymmetricKey, identifier: KeyIdentifier) -> CardResult<Void> {
-    let primary = keychain.storeSymmetricKey(key, tag: identifier.primaryTag)
-    guard case .success = primary else { return primary }
-    _ = keychain.storeSymmetricKey(key, tag: identifier.legacyKeychainTag)
-    return .success(())
-  }
-
-  private func storePrivateKey(_ key: P256.Signing.PrivateKey, identifier: KeyIdentifier) -> CardResult<Void> {
-    let primary = keychain.storePrivateKey(key, tag: identifier.primaryTag)
-    guard case .success = primary else { return primary }
-    _ = keychain.storePrivateKey(key, tag: identifier.legacyKeychainTag)
-    return .success(())
-  }
-}
-
-// MARK: - Supporting Types
-
-struct PublicKeyBundle: Codable {
-  let signingPublicKey: Data
-  let keyId: String
-  let createdAt: Date
-  let expiresAt: Date
-
-  var isExpired: Bool {
-    return expiresAt < Date()
-  }
-
-  var isValid: Bool {
-    return !isExpired && signingPublicKey.count == 64  // P256 public key size
-  }
-}
-
-// MARK: - Keychain Manager
-
-private class KeychainManager {
-
-  func storeSymmetricKey(_ key: SymmetricKey, tag: String) -> CardResult<Void> {
-    let keyData = key.withUnsafeBytes { Data($0) }
-
-    let query: [String: Any] = [
-      kSecClass as String: kSecClassGenericPassword,
-      kSecAttrAccount as String: tag,
-      kSecValueData as String: keyData,
-      kSecAttrAccessible as String: kSecAttrAccessibleWhenUnlocked,
-      kSecAttrSynchronizable as String: kCFBooleanTrue as Any,
-    ]
-
-    // Delete existing key first
-    _ = deleteKey(tag: tag)
-
-    let status = SecItemAdd(query as CFDictionary, nil)
-
-    if status == errSecSuccess {
-      return .success(())
-    } else {
-      return .failure(.encryptionError("Failed to store symmetric key: \(status)"))
-    }
-  }
-
-  func retrieveSymmetricKey(tag: String) -> CardResult<SymmetricKey> {
-    let query: [String: Any] = [
-      kSecClass as String: kSecClassGenericPassword,
-      kSecAttrAccount as String: tag,
-      kSecReturnData as String: true,
-      kSecMatchLimit as String: kSecMatchLimitOne,
-    ]
-
-    var result: AnyObject?
-    let status = SecItemCopyMatching(query as CFDictionary, &result)
-
-    if status == errSecSuccess, let keyData = result as? Data {
-      let key = SymmetricKey(data: keyData)
-      return .success(key)
-    } else {
-      return .failure(.encryptionError("Failed to retrieve symmetric key: \(status)"))
-    }
-  }
-
-  func storePrivateKey(_ key: P256.Signing.PrivateKey, tag: String) -> CardResult<Void> {
-    let keyData = key.rawRepresentation
-
-    let query: [String: Any] = [
-      kSecClass as String: kSecClassGenericPassword,
-      kSecAttrAccount as String: tag,
-      kSecValueData as String: keyData,
-      kSecAttrAccessible as String: kSecAttrAccessibleWhenUnlocked,
-      kSecAttrSynchronizable as String: kCFBooleanTrue as Any,
-    ]
-
-    // Delete existing key first
-    _ = deleteKey(tag: tag)
-
-    let status = SecItemAdd(query as CFDictionary, nil)
-
-    if status == errSecSuccess {
-      return .success(())
-    } else {
-      return .failure(.encryptionError("Failed to store private key: \(status)"))
-    }
-  }
-
-  func retrievePrivateKey(tag: String) -> CardResult<P256.Signing.PrivateKey> {
-    let query: [String: Any] = [
-      kSecClass as String: kSecClassGenericPassword,
-      kSecAttrAccount as String: tag,
-      kSecReturnData as String: true,
-      kSecMatchLimit as String: kSecMatchLimitOne,
-    ]
-
-    var result: AnyObject?
-    let status = SecItemCopyMatching(query as CFDictionary, &result)
-
-    if status == errSecSuccess, let keyData = result as? Data {
-      do {
-        let key = try P256.Signing.PrivateKey(rawRepresentation: keyData)
-        return .success(key)
-      } catch {
-        return .failure(.encryptionError("Failed to reconstruct private key: \(error.localizedDescription)"))
-      }
-    } else {
-      return .failure(.encryptionError("Failed to retrieve private key: \(status)"))
-    }
-  }
-
-  func deleteKey(tag: String) -> CardResult<Void> {
-    let query: [String: Any] = [
-      kSecClass as String: kSecClassGenericPassword,
-      kSecAttrAccount as String: tag,
-    ]
-
-    let status = SecItemDelete(query as CFDictionary)
-
-    if status == errSecSuccess || status == errSecItemNotFound {
-      return .success(())
-    } else {
-      return .failure(.encryptionError("Failed to delete key: \(status)"))
-    }
-  }
 }
