+++ solidarity/Services/Vault/InactivityMonitorService.swift
@@ -24,6 +24,12 @@ final class InactivityMonitorService: ObservableObject {
     @Published private(set) var pendingUnlocks: [PendingUnlock] = []
     @Published private(set) var isMonitoring = true
 
+    // MARK: - Internal Properties
+
+    let vault = SovereignVaultService.shared
+    let notificationCenter = UNUserNotificationCenter.current()
+    let warningDays = [7, 3, 1]
+
     // MARK: - Private Properties
 
     private let userDefaults = UserDefaults.standard
@@ -31,19 +37,11 @@ final class InactivityMonitorService: ObservableObject {
     private let historyKey = "com.solidarity.vault.activityHistory"
     private var cancellables = Set<AnyCancellable>()
     private var checkTimer: Timer?
-
-    private let vault = SovereignVaultService.shared
-    private let notificationCenter = UNUserNotificationCenter.current()
-
-    // MARK: - Configuration
-
-    private let warningDays = [7, 3, 1] // Days before unlock to warn
-    private let minimumCheckInterval: TimeInterval = 3600 // 1 hour
+    private let minimumCheckInterval: TimeInterval = 3600
 
     // MARK: - Initialization
 
     private init() {
-        // Load last activity date or initialize to now
         if let savedDate = userDefaults.object(forKey: activityKey) as? Date {
             lastActivityDate = savedDate
         } else {
@@ -64,29 +62,21 @@ final class InactivityMonitorService: ObservableObject {
 
     // MARK: - Public API
 
-    /// Record user activity (call on meaningful interactions)
     func recordActivity() {
         lastActivityDate = Date()
         daysSinceActivity = 0
         saveActivity()
-
-        // Cancel any pending unlock notifications
         cancelPendingNotifications()
-
-        // Reset pending unlocks
         pendingUnlocks.removeAll()
-
         print("[InactivityMonitor] Activity recorded at \(lastActivityDate)")
     }
 
-    /// Perform a check for inactivity-based unlocks
     func performCheck() {
         updateDaysSinceActivity()
         checkForPendingUnlocks()
         scheduleWarningNotifications()
     }
 
-    /// Get items that will unlock soon due to inactivity
     func itemsWithUpcomingUnlock(within days: Int) -> [VaultItem] {
         vault.items.filter { item in
             guard let config = item.timeLockConfig,
@@ -100,7 +90,6 @@ final class InactivityMonitorService: ObservableObject {
         }
     }
 
-    /// Manually trigger the recovery process for an item
     func initiateRecoveryProcess(for itemId: UUID) async throws {
         guard let item = vault.items.first(where: { $0.id == itemId }),
               let config = item.timeLockConfig,
@@ -108,21 +97,17 @@ final class InactivityMonitorService: ObservableObject {
             throw InactivityError.itemNotConfigured
         }
 
-        // Check if inactivity threshold has been met
         if let inactivityDays = config.inactivityDays,
            daysSinceActivity >= inactivityDays {
-            // Trigger the unlock flow
             try await processAutoUnlock(item: item)
         } else if let unlockDate = config.unlockDate,
                   Date() >= unlockDate {
-            // Fixed date unlock
             try await processAutoUnlock(item: item)
         } else {
             throw InactivityError.unlockConditionsNotMet
         }
     }
 
-    /// Configure inactivity monitoring for an item
     func configureMonitoring(
         for itemId: UUID,
         inactivityDays: Int,
@@ -136,7 +121,6 @@ final class InactivityMonitorService: ObservableObject {
             witnessContactIds: witnessContactIds
         )
 
-        // If there's a beneficiary, create and distribute key shards
         if let beneficiaryId = beneficiaryContactId {
             let shards = try await createKeyShards(
                 for: itemId,
@@ -150,7 +134,6 @@ final class InactivityMonitorService: ObservableObject {
         try await vault.updateTimeLock(itemId, config: config)
     }
 
-    /// Get status summary
     var statusSummary: ActivityStatus {
         ActivityStatus(
             lastActivity: lastActivityDate,
@@ -164,14 +147,12 @@ final class InactivityMonitorService: ObservableObject {
     // MARK: - Private Methods
 
     private func setupActivityObservers() {
-        // Observe app becoming active
         NotificationCenter.default.publisher(for: UIApplication.didBecomeActiveNotification)
             .sink { [weak self] _ in
                 self?.performCheck()
             }
             .store(in: &cancellables)
 
-        // Observe significant time changes
         NotificationCenter.default.publisher(for: UIApplication.significantTimeChangeNotification)
             .sink { [weak self] _ in
                 self?.updateDaysSinceActivity()
@@ -197,7 +178,6 @@ final class InactivityMonitorService: ObservableObject {
     private func saveActivity() {
         userDefaults.set(lastActivityDate, forKey: activityKey)
 
-        // Save to history (keep last 90 days)
         var history = userDefaults.array(forKey: historyKey) as? [Date] ?? []
         history.append(lastActivityDate)
         let ninetyDaysAgo = Date().addingTimeInterval(-90 * 24 * 60 * 60)
@@ -212,12 +192,10 @@ final class InactivityMonitorService: ObservableObject {
             guard let config = item.timeLockConfig,
                   config.enabled else { continue }
 
-            // Check inactivity-based unlock
             if let inactivityDays = config.inactivityDays {
                 let daysUntilUnlock = inactivityDays - daysSinceActivity
 
                 if daysUntilUnlock <= 0 {
-                    // Should unlock now
                     Task {
                         try? await processAutoUnlock(item: item)
                     }
@@ -232,7 +210,6 @@ final class InactivityMonitorService: ObservableObject {
                 }
             }
 
-            // Check fixed date unlock
             if let unlockDate = config.unlockDate {
                 if Date() >= unlockDate {
                     Task {
@@ -254,185 +231,5 @@ final class InactivityMonitorService: ObservableObject {
         }
 
         pendingUnlocks = newPendingUnlocks
-    }
-
-    private func processAutoUnlock(item: VaultItem) async throws {
-        guard var config = item.timeLockConfig else { return }
-
-        // Update status
-        config.status = .unlocked
-        try await vault.updateTimeLock(item.id, config: config)
-
-        // Notify beneficiary if configured
-        if let beneficiaryId = config.beneficiaryContactId {
-            await notifyBeneficiary(beneficiaryId, about: item)
-        }
-
-        // Log the unlock event
-        print("[InactivityMonitor] Auto-unlocked item: \(item.name)")
-    }
-
-    private func createKeyShards(
-        for itemId: UUID,
-        beneficiary: UUID,
-        witnesses: [UUID]
-    ) async throws -> [EncryptedKeyShard] {
-        // Get the item's encryption key
-        let itemKey = try await getItemEncryptionKey(itemId)
-
-        // Split into shards (beneficiary + witnesses)
-        let totalParties = 1 + witnesses.count
-        let threshold = max(2, (totalParties / 2) + 1)
-
-        let shares = try ShamirSecretSharing.split(
-            secret: itemKey,
-            threshold: threshold,
-            totalShares: totalParties
-        )
-
-        // Create encrypted shards for each party
-        var shards: [EncryptedKeyShard] = []
-
-        // Beneficiary gets first shard
-        shards.append(EncryptedKeyShard(
-            shardIndex: 0,
-            encryptedData: shares[0].value, // In production, encrypt with beneficiary's public key
-            recipientContactId: beneficiary
-        ))
-
-        // Witnesses get remaining shards
-        for (index, witnessId) in witnesses.enumerated() {
-            shards.append(EncryptedKeyShard(
-                shardIndex: index + 1,
-                encryptedData: shares[index + 1].value,
-                recipientContactId: witnessId
-            ))
-        }
-
-        return shards
-    }
-
-    private func getItemEncryptionKey(_ itemId: UUID) async throws -> Data {
-        // In production, retrieve the actual item key
-        // For now, generate a placeholder
-        let randomBytes = (0..<32).map { _ in UInt8.random(in: 0...255) }
-        return Data(randomBytes)
-    }
-
-    private func notifyBeneficiary(_ beneficiaryId: UUID, about item: VaultItem) async {
-        // In production, send a notification to the beneficiary
-        // This could be via push notification, email, or the app's messaging system
-        print("[InactivityMonitor] Would notify beneficiary \(beneficiaryId) about unlocked item: \(item.name)")
-    }
-
-    // MARK: - Notifications
-
-    private func requestNotificationPermission() {
-        notificationCenter.requestAuthorization(options: [.alert, .badge, .sound]) { granted, _ in
-            if granted {
-                print("[InactivityMonitor] Notification permission granted")
-            }
-        }
-    }
-
-    private func scheduleWarningNotifications() {
-        for pending in pendingUnlocks {
-            scheduleNotification(for: pending)
-        }
-    }
-
-    private func scheduleNotification(for pending: PendingUnlock) {
-        let content = UNMutableNotificationContent()
-        content.title = String(localized: "Vault Item Will Unlock Soon")
-        let bodyFormat = String(localized: "\"%@\" will unlock in %lld day(s). Open the app to record activity.")
-        content.body = String(format: bodyFormat, locale: Locale.current, pending.itemName, pending.daysUntilUnlock)
-        content.sound = .default
-        content.categoryIdentifier = "VAULT_UNLOCK_WARNING"
-
-        let trigger = UNTimeIntervalNotificationTrigger(
-            timeInterval: 60, // Show soon (for testing; in production, space these out)
-            repeats: false
-        )
-
-        let request = UNNotificationRequest(
-            identifier: "vault-unlock-\(pending.itemId.uuidString)",
-            content: content,
-            trigger: trigger
-        )
-
-        notificationCenter.add(request)
-    }
-
-    private func cancelPendingNotifications() {
-        let identifiers = pendingUnlocks.map { "vault-unlock-\($0.itemId.uuidString)" }
-        notificationCenter.removePendingNotificationRequests(withIdentifiers: identifiers)
-    }
-}
-
-// MARK: - Supporting Types
-
-struct PendingUnlock: Identifiable {
-    var id: UUID { itemId }
-    let itemId: UUID
-    let itemName: String
-    let daysUntilUnlock: Int
-    let unlockType: UnlockType
-    let beneficiaryId: UUID?
-
-    enum UnlockType {
-        case inactivity
-        case fixedDate(Date)
-
-        var description: String {
-            switch self {
-            case .inactivity:
-                return String(localized: "Inactivity unlock")
-            case .fixedDate(let date):
-                let format = String(localized: "Scheduled for %@")
-                return String(format: format, locale: Locale.current, date.formatted(date: .abbreviated, time: .omitted))
-            }
-        }
-    }
-}
-
-struct ActivityStatus {
-    let lastActivity: Date
-    let daysSinceActivity: Int
-    let monitoredItemCount: Int
-    let pendingUnlockCount: Int
-    let isMonitoring: Bool
-
-    var formattedLastActivity: String {
-        lastActivity.formatted(date: .abbreviated, time: .shortened)
-    }
-
-    var statusMessage: String {
-        if daysSinceActivity == 0 {
-            return String(localized: "Active today")
-        } else if daysSinceActivity == 1 {
-            return String(localized: "Last active yesterday")
-        } else {
-            let format = String(localized: "Last active %lld days ago")
-            return String(format: format, locale: Locale.current, daysSinceActivity)
-        }
-    }
-}
-
-// MARK: - Errors
-
-enum InactivityError: LocalizedError {
-    case itemNotConfigured
-    case unlockConditionsNotMet
-    case shardCreationFailed
-
-    var errorDescription: String? {
-        switch self {
-        case .itemNotConfigured:
-            return String(localized: "Item is not configured for inactivity monitoring")
-        case .unlockConditionsNotMet:
-            return String(localized: "Unlock conditions have not been met")
-        case .shardCreationFailed:
-            return String(localized: "Failed to create key shards")
-        }
     }
 }
