+++ solidarity/Services/ZK/ProofGenerationManager.swift
@@ -12,7 +12,7 @@ import Foundation
 class ProofGenerationManager {
   static let shared = ProofGenerationManager()
 
-  private let keyManager = KeyManager.shared
+  let keyManager = KeyManager.shared
   private let domainVerificationManager = DomainVerificationManager.shared
 
   private init() {}
@@ -91,286 +91,9 @@ class ProofGenerationManager {
     }
   }
 
-  /// Verify selective disclosure proof
-  func verifySelectiveDisclosureProof(
-    _ proof: SelectiveDisclosureProof,
-    expectedBusinessCardId: String?
-  ) -> CardResult<ProofVerificationResult> {
+  // MARK: - Internal Helpers
 
-    // Check expiration
-    if proof.expiresAt < Date() {
-      return .success(
-        ProofVerificationResult(
-          isValid: false,
-          reason: "Proof has expired",
-          verifiedAt: Date()
-        )
-      )
-    }
-
-    // Check business card ID if provided
-    if let expectedId = expectedBusinessCardId, proof.businessCardId != expectedId {
-      return .success(
-        ProofVerificationResult(
-          isValid: false,
-          reason: "Business card ID mismatch",
-          verifiedAt: Date()
-        )
-      )
-    }
-
-    // Verify signature using the same deterministic payload as signing
-    let proofData = canonicalProofSigningData(
-      businessCardId: proof.businessCardId,
-      selectedFields: Set(proof.disclosedFields.keys),
-      recipientId: proof.recipientId,
-      timestamp: proof.createdAt
-    )
-
-    guard let proofDataToVerify = proofData else {
-      return .success(
-        ProofVerificationResult(
-          isValid: false,
-          reason: "Failed to reconstruct proof data",
-          verifiedAt: Date()
-        )
-      )
-    }
-
-    let signatureResult = verifyProofSignature(proofDataToVerify, signature: proof.signature)
-
-    switch signatureResult {
-    case .success(let isValidSignature):
-      if isValidSignature {
-        return .success(
-          ProofVerificationResult(
-            isValid: true,
-            reason: "Proof is valid",
-            verifiedAt: Date()
-          )
-        )
-      } else {
-        return .success(
-          ProofVerificationResult(
-            isValid: false,
-            reason: "Invalid signature",
-            verifiedAt: Date()
-          )
-        )
-      }
-
-    case .failure(let error):
-      return .failure(error)
-    }
-  }
-
-  /// Generate attribute proof (e.g., "has skill X" without revealing other skills)
-  func generateAttributeProof(
-    businessCard: BusinessCard,
-    attribute: AttributeType,
-    value: String
-  ) -> CardResult<AttributeProof> {
-
-    let keyResult = keyManager.getMasterKey()
-
-    switch keyResult {
-    case .success(let masterKey):
-      // Check if attribute exists
-      let hasAttribute = checkAttributeExists(
-        in: businessCard,
-        attribute: attribute,
-        value: value
-      )
-
-      if !hasAttribute {
-        return .failure(.validationError("Attribute not found in business card"))
-      }
-
-      // Generate attribute commitment
-      let attributeData = Data("\(attribute.rawValue):\(value)".utf8)
-      let cardIdData = Data(businessCard.id.uuidString.utf8)
-
-      let commitment = SHA256.hash(data: attributeData + cardIdData + masterKey.withUnsafeBytes { Data($0) })
-
-      // Generate proof signature
-      let proofData = try? JSONEncoder()
-        .encode(
-          AttributeProofData(
-            businessCardId: businessCard.id.uuidString,
-            attributeType: attribute,
-            hasAttribute: true,
-            timestamp: Date()
-          )
-        )
-
-      guard let proofDataToSign = proofData else {
-        return .failure(.encryptionError("Failed to encode attribute proof data"))
-      }
-
-      let signatureResult = signProofData(proofDataToSign)
-
-      switch signatureResult {
-      case .success(let signature):
-        return .success(
-          AttributeProof(
-            proofId: UUID().uuidString,
-            businessCardId: businessCard.id.uuidString,
-            attributeType: attribute,
-            commitment: Data(commitment),
-            signature: signature,
-            createdAt: Date(),
-            expiresAt: Calendar.current.date(byAdding: .hour, value: 12, to: Date()) ?? Date()
-          )
-        )
-
-      case .failure(let error):
-        return .failure(error)
-      }
-
-    case .failure(let error):
-      return .failure(error)
-    }
-  }
-
-  /// Verify attribute proof
-  func verifyAttributeProof(
-    _ proof: AttributeProof,
-    expectedAttribute: AttributeType,
-    expectedValue: String
-  ) -> CardResult<ProofVerificationResult> {
-
-    // Check expiration
-    if proof.expiresAt < Date() {
-      return .success(
-        ProofVerificationResult(
-          isValid: false,
-          reason: "Proof has expired",
-          verifiedAt: Date()
-        )
-      )
-    }
-
-    // Check attribute type
-    if proof.attributeType != expectedAttribute {
-      return .success(
-        ProofVerificationResult(
-          isValid: false,
-          reason: "Attribute type mismatch",
-          verifiedAt: Date()
-        )
-      )
-    }
-
-    // Verify signature
-    let proofData = try? JSONEncoder()
-      .encode(
-        AttributeProofData(
-          businessCardId: proof.businessCardId,
-          attributeType: proof.attributeType,
-          hasAttribute: true,
-          timestamp: proof.createdAt
-        )
-      )
-
-    guard let proofDataToVerify = proofData else {
-      return .success(
-        ProofVerificationResult(
-          isValid: false,
-          reason: "Failed to reconstruct proof data",
-          verifiedAt: Date()
-        )
-      )
-    }
-
-    let signatureResult = verifyProofSignature(proofDataToVerify, signature: proof.signature)
-
-    switch signatureResult {
-    case .success(let isValidSignature):
-      return .success(
-        ProofVerificationResult(
-          isValid: isValidSignature,
-          reason: isValidSignature ? "Proof is valid" : "Invalid signature",
-          verifiedAt: Date()
-        )
-      )
-
-    case .failure(let error):
-      return .failure(error)
-    }
-  }
-
-  /// Generate range proof (e.g., "experience > 5 years" without revealing exact years)
-  func generateRangeProof(
-    businessCard: BusinessCard,
-    attribute: AttributeType,
-    range: ClosedRange<Int>
-  ) -> CardResult<RangeProof> {
-
-    let actualValue = getAttributeNumericValue(from: businessCard, attribute: attribute)
-
-    guard let value = actualValue else {
-      return .failure(.validationError("Attribute not found or not numeric"))
-    }
-
-    let isInRange = range.contains(value)
-
-    let keyResult = keyManager.getMasterKey()
-
-    switch keyResult {
-    case .success(let masterKey):
-      // Generate range commitment
-      let rangeData = Data("\(attribute.rawValue):\(range.lowerBound)-\(range.upperBound)".utf8)
-      let cardIdData = Data(businessCard.id.uuidString.utf8)
-      let resultData = Data((isInRange ? "true" : "false").utf8)
-
-      let commitment = SHA256.hash(data: rangeData + cardIdData + resultData + masterKey.withUnsafeBytes { Data($0) })
-
-      // Generate proof signature
-      let proofData = try? JSONEncoder()
-        .encode(
-          RangeProofData(
-            businessCardId: businessCard.id.uuidString,
-            attributeType: attribute,
-            range: range,
-            isInRange: isInRange,
-            timestamp: Date()
-          )
-        )
-
-      guard let proofDataToSign = proofData else {
-        return .failure(.encryptionError("Failed to encode range proof data"))
-      }
-
-      let signatureResult = signProofData(proofDataToSign)
-
-      switch signatureResult {
-      case .success(let signature):
-        return .success(
-          RangeProof(
-            proofId: UUID().uuidString,
-            businessCardId: businessCard.id.uuidString,
-            attributeType: attribute,
-            range: range,
-            isInRange: isInRange,
-            commitment: Data(commitment),
-            signature: signature,
-            createdAt: Date(),
-            expiresAt: Calendar.current.date(byAdding: .hour, value: 6, to: Date()) ?? Date()
-          )
-        )
-
-      case .failure(let error):
-        return .failure(error)
-      }
-
-    case .failure(let error):
-      return .failure(error)
-    }
-  }
-
-  // MARK: - Private Methods
-
-  private func getFieldValue(from businessCard: BusinessCard, field: BusinessCardField) -> String {
+  func getFieldValue(from businessCard: BusinessCard, field: BusinessCardField) -> String {
     switch field {
     case .name: return businessCard.name
     case .title: return businessCard.title ?? ""
@@ -387,7 +110,7 @@ class ProofGenerationManager {
   /// Build a deterministic signing payload for selective disclosure proofs.
   /// We avoid JSON encoder ordering issues by constructing a canonical string:
   /// businessCardId|field1,field2,...|recipientId|timestampSeconds
-  private func canonicalProofSigningData(
+  func canonicalProofSigningData(
     businessCardId: String,
     selectedFields: Set<BusinessCardField>,
     recipientId: String?,
@@ -400,6 +123,43 @@ class ProofGenerationManager {
     let canonical = [businessCardId, fieldNames.joined(separator: ","), recipient, ts].joined(separator: "|")
     return canonical.data(using: .utf8)
   }
+
+  func signProofData(_ data: Data) -> CardResult<Data> {
+    let keyResult = keyManager.getSigningKeyPair()
+
+    switch keyResult {
+    case .success(let keyPair):
+      do {
+        let signature = try keyPair.privateKey.signature(for: data)
+        return .success(signature.rawRepresentation)
+      } catch {
+        return .failure(.encryptionError("Failed to sign proof data: \(error.localizedDescription)"))
+      }
+
+    case .failure(let error):
+      return .failure(error)
+    }
+  }
+
+  func verifyProofSignature(_ data: Data, signature: Data) -> CardResult<Bool> {
+    let keyResult = keyManager.getSigningKeyPair()
+
+    switch keyResult {
+    case .success(let keyPair):
+      do {
+        let ecdsaSignature = try P256.Signing.ECDSASignature(rawRepresentation: signature)
+        let isValid = keyPair.publicKey.isValidSignature(ecdsaSignature, for: data)
+        return .success(isValid)
+      } catch {
+        return .success(false)
+      }
+
+    case .failure:
+      return .success(false)
+    }
+  }
+
+  // MARK: - Private Methods
 
   private func generateFieldCommitment(
     field: BusinessCardField,
@@ -419,77 +179,5 @@ class ProofGenerationManager {
 
     let commitment = SHA256.hash(data: fieldData + recipientData + masterKey.withUnsafeBytes { Data($0) })
     return Data(commitment)
-  }
-
-  private func signProofData(_ data: Data) -> CardResult<Data> {
-    let keyResult = keyManager.getSigningKeyPair()
-
-    switch keyResult {
-    case .success(let keyPair):
-      do {
-        let signature = try keyPair.privateKey.signature(for: data)
-        return .success(signature.rawRepresentation)
-      } catch {
-        return .failure(.encryptionError("Failed to sign proof data: \(error.localizedDescription)"))
-      }
-
-    case .failure(let error):
-      return .failure(error)
-    }
-  }
-
-  private func verifyProofSignature(_ data: Data, signature: Data) -> CardResult<Bool> {
-    let keyResult = keyManager.getSigningKeyPair()
-
-    switch keyResult {
-    case .success(let keyPair):
-      do {
-        let ecdsaSignature = try P256.Signing.ECDSASignature(rawRepresentation: signature)
-        let isValid = keyPair.publicKey.isValidSignature(ecdsaSignature, for: data)
-        return .success(isValid)
-      } catch {
-        return .success(false)
-      }
-
-    case .failure:
-      return .success(false)
-    }
-  }
-
-  private func checkAttributeExists(
-    in businessCard: BusinessCard,
-    attribute: AttributeType,
-    value: String
-  ) -> Bool {
-    // Compare by first 3 characters to avoid leaking full data and to simplify checks
-    let target = String(value.prefix(3)).lowercased()
-    switch attribute {
-    case .skill:
-      return businessCard.skills.contains { String($0.name.prefix(3)).lowercased() == target }
-    case .company:
-      return String((businessCard.company ?? "").prefix(3)).lowercased() == target
-    case .title:
-      return String((businessCard.title ?? "").prefix(3)).lowercased() == target
-    case .domain:
-      if let email = businessCard.email {
-        let domain = email.components(separatedBy: "@").last?.lowercased()
-        return String((domain ?? "").prefix(3)).lowercased() == target
-      }
-      return false
-    }
-  }
-
-  private func getAttributeNumericValue(
-    from businessCard: BusinessCard,
-    attribute: AttributeType
-  ) -> Int? {
-    switch attribute {
-    case .skill:
-      // Return number of skills
-      return businessCard.skills.count
-    case .company, .title, .domain:
-      // These are not numeric attributes
-      return nil
-    }
   }
 }
